Fixed: Catch an invalid query at compile time
When sending a query off to Verovio, it can raise an invalid query if the input cannot be interpreted as music notation. This is not possible to check until we send it to Verovio.

However, this exception was not caught, meaning the server crashed on single-note input.

This commit fixes this issue by moving the compile step into the try/except block.

Refs https://github.com/rism-digital/rism-online-issues/issues/102