Re-work of configuration file
Removes the different modes for select facets; this was too complicated to manage.

Adds a new facet type of 'query' that will show a query box tied to a full-text query.

Improves the documentation for this.

Improves the method by which facet values are translated. Rather than passing a set number of values, this will now take the values from Solr directly, and simply prepend the correct section from the 'translation_prefix' field.