Fixed: Refactor checks on source sections
The previous logic would render the full section before returning, and then we would check the contents
of the returned value to determine whether this section should be in the output or not (returning None).

Now we check some proxy values first to determine whether we should actually render the value, and return None
straight away if the answer was going to be None anyway.