Fixed: Hide composer on single-source view
When viewing a list of incipits on the source, the composer is now hidden
since it will always be the same composer. The composer is shown when the
incipit record is shown on its own, because otherwise the user would have to
view the source record to determine the composer.

Fixes rism-digital/rism-online-issues#164