Fixed: Extensive rework of title rendering
Most of the changes in this commit have to do with moving the translations and formatters to
the shared helpers. However, the titles for sources are now more completely translated, more
consistent, and more aligned with the OPAC's title format.

This change was triggered by a request that the keys were not available in the titles. This is also fixed.

Fixes rism-digital/rism-online-issues#220