Fixed: Composite sources
Composite sources are now served correctly. This commit adds a few things:

 - When a composite source is detected, Solr will be queried for both sources and holdings
 - If a holding record is retrieved, an extra Solr lookup will be performed to fetch the corresponding source record.
 - Fixes an issue where if a composite record consisted of only holdings, then the 'contents' tab would not appear
 - Changes the 'sort' block in the response to deliver the 'default' sort key

Also adds an explicit sort option for the source "contents" search that allows for displaying the items in source order. Reworks the 'sorting' parameter to accurately decide on the actual default sorting behaviour for Solr.

Fixes rism-digital/rism-online-issues#230