Fixed: Update context handling
The context handling is now done centrally in the request handler, where we can
embed the correct context prior to sending it to be serialized.

For non-JSON-LD responses, the expanded context should be provided, since it
removes the need for a further lookup when it goes through the RDFlib serializer.