New: Use embeddable record view for search engines
A bug with the server configuration meant that Google wasn't able to load the JavaScript page correctly. Since this was fixed, the full record view can be seen by Google.

We will keep this in the export server for now, since we still need an HTML page for those bots that do not execute JavaScript.