Fixed: Incorrect validation on incipit queries
The new query validation was running on incipit queries and failing because incipit queries inject raw Solr fields into the q parameter, and then the query validator is run. Because the fields are not in the list of acceptable replacements (UI -> Solr) they would be marked as an invalid query.

This change uses a new parameter in the small-asc query validator that is used to specify "raw" fields that will bypass the validation.